test: 우동 참가자 수 조회 쿼리 udongRepository로 이동(쿼리에서 coalesce하기 위함)
