+++ src/main/java/com/hyun/udong/udong/infrastructure/repository/ParticipantRepository.java
@@ -3,5 +3,5 @@ package com.hyun.udong.udong.infrastructure.repository;
 import com.hyun.udong.udong.domain.Participant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipantRepository extends JpaRepository<Participant, Long>, ParticipantRepositoryCustom {
+public interface ParticipantRepository extends JpaRepository<Participant, Long> {
 }
