--- src/main/java/com/hyun/udong/udong/infrastructure/repository/ParticipantRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.hyun.udong.udong.infrastructure.repository;
-
-import com.hyun.udong.udong.presentation.dto.response.ParticipantCountResponse;
-
-import java.util.List;
-
-public interface ParticipantRepositoryCustom {
-    List<ParticipantCountResponse> countParticipantsByUdongIds(List<Long> udongIds);
-}
