+++ src/main/java/com/hyun/udong/udong/infrastructure/repository/UdongRepositoryImpl.java
@@ -2,6 +2,8 @@ package com.hyun.udong.udong.infrastructure.repository;
 
 import com.hyun.udong.udong.domain.Udong;
 import com.hyun.udong.udong.presentation.dto.request.FindUdongsCondition;
+import com.hyun.udong.udong.presentation.dto.response.ParticipantCountResponse;
+import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -14,6 +16,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.hyun.udong.common.util.querydsl.QueryDslUtils.*;
+import static com.hyun.udong.udong.domain.QParticipant.participant;
 import static com.hyun.udong.udong.domain.QTravelCity.travelCity;
 import static com.hyun.udong.udong.domain.QUdong.udong;
 
@@ -42,6 +45,20 @@ public class UdongRepositoryImpl implements UdongRepositoryCustom {
                 .fetchOne()).orElse(0L);
 
         return new PageImpl<>(udongs, pageable, total);
+    }
+
+    @Override
+    public List<ParticipantCountResponse> countParticipantsByUdongIds(List<Long> udongIds) {
+        return queryFactory
+                .select(Projections.constructor(ParticipantCountResponse.class,
+                        udong.id,
+                        participant.count().coalesce(0L)
+                ))
+                .from(udong)
+                .leftJoin(participant).on(participant.udong.eq(udong))
+                .where(udong.id.in(udongIds))
+                .groupBy(udong.id)
+                .fetch();
     }
 
     private BooleanExpression createWhere(FindUdongsCondition filter) {
