+++ src/main/java/com/hyun/udong/udong/presentation/dto/response/ParticipantCountResponse.java
@@ -1,4 +1,4 @@
 package com.hyun.udong.udong.presentation.dto.response;
 
-public record ParticipantCountResponse(Long udongId, int participantCount) {
+public record ParticipantCountResponse(Long udongId, Long participantCount) {
 }
